requirements.txt
asgiref==3.11.1
django-axes==8.3.1
cloudinary==1.44.1
django-redis==6.0.0
redis==7.2.1
colorama==0.4.6
dj-database-url==0.5.0
django-allauth==0.57.2
django-cloudinary-storage==0.3.0
Django==4.2.28
gunicorn==20.1.0
psycopg2-binary==2.9.9
requests==2.32.3
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
whitenoise==6.7.0

channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-csp==3.8
bleach==6.2.0
phonenumbers==9.0.25
